Tweeter Android App/tweeter-main/app/src/main/java/edu/byu/cs/tweeter/client/model/service/service/StatusService.java
package edu.byu.cs.tweeter.client.model.service.service;

import edu.byu.cs.tweeter.client.cache.Cache;
import edu.byu.cs.tweeter.client.model.service.backgroundTask.task.GetFeedTask;
import edu.byu.cs.tweeter.client.model.service.backgroundTask.task.GetStoryTask;
import edu.byu.cs.tweeter.client.model.service.backgroundTask.task.PostStatusTask;
import edu.byu.cs.tweeter.client.model.service.backgroundTask.handler.GetFeedHandler;
import edu.byu.cs.tweeter.client.model.service.backgroundTask.handler.GetStoryHandler;
import edu.byu.cs.tweeter.client.model.service.backgroundTask.handler.PostStatusHandler;
import edu.byu.cs.tweeter.client.model.service.observer.PageObserver;
import edu.byu.cs.tweeter.client.model.service.observer.StatusObserver;
import edu.byu.cs.tweeter.model.domain.Status;
import edu.byu.cs.tweeter.model.domain.User;

public class StatusService extends BaseService {
    public void loadMoreFeed(User user, int pageSize, Status lastStatus, PageObserver<Status> observer) {
        GetFeedTask getFeedTask = new GetFeedTask(Cache.getInstance().getCurrUserAuthToken(),
                user, pageSize, lastStatus, new GetFeedHandler(observer));
        execute(getFeedTask);
    }

    public void loadMoreStory(User user, int pageSize, Status lastStatus, PageObserver<Status> observer) {
        GetStoryTask getStoryTask = new GetStoryTask(Cache.getInstance().getCurrUserAuthToken(),
                user, pageSize, lastStatus, new GetStoryHandler(observer));
        execute(getStoryTask);
    }

    public void postStatus(Status newStatus, StatusObserver observer) {
        PostStatusTask statusTask = new PostStatusTask(Cache.getInstance().getCurrUserAuthToken(),
                newStatus, new PostStatusHandler(observer));
        execute(statusTask);
    }
}
